add industry id in navigation

+++ app/src/main/java/ru/practicum/android/diploma/ui/navigation/util/AppGraph.kt
@@ -5,7 +5,7 @@ import androidx.navigation.NavGraphBuilder
 fun NavGraphBuilder.appGraph(
     navigateToVacancy: (String) -> Unit,
     onBackClick: () -> Unit,
-    navigateToFilterIndustry: () -> Unit,
+    navigateToFilterIndustry: (Int) -> Unit,
     navigateToFilterSettings: () -> Unit,
     onFinishFilterIndustry: () -> Unit
 ) {
